chore(claude-provider): split event-translation helpers out so the file fits the 500-line gate
The provider file landed at 551 lines after #131 (Claude tool bridge +
streaming-mode prompt input), which trips `scripts/check-file-lines.mjs`
in CI and is currently blocking #138, #139, and #140.

Fix: extract the SDK-message → StreamEvent translation layer into its
own module (`_claude_events.py`):

  - _events_from_message
  - _events_from_assistant
  - _tool_result_event
  - _tool_result_to_text
  - _error_event

Re-exported from `claude_provider` so existing import sites and tests
(`tests/test_claude_provider.py` imports them as private names from
the provider module) keep working without churn. The split is internal:
no public API change, no behaviour change.

Drops unused `claude_agent_sdk` imports from `claude_provider` since
the message types now live in `_claude_events`.

Result:
  - claude_provider.py: 551 → 464 lines (well under the 500 gate)
  - new _claude_events.py: 135 lines

55/55 backend tests pass locally.
